Update to send address in chat message
Be warned that this does not always work behind a firewall and you will
get a local network address, and Linphone will not be able to connect. I
am leaving this code in place because, as I understand the issue, this
will work on some systems/routers, so it will be useful to some.